+++ app/lib/shared/versions.dart
@@ -24,10 +24,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const _acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2025.05.21',
+  '2025.06.02',
   // Fallback runtime versions.
+  '2025.05.21',
   '2025.05.15',
-  '2025.05.13',
 ];
 
 /// Sets the current runtime versions.
@@ -62,8 +62,8 @@ bool shouldGCVersion(String version) =>
 
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '3.8.0';
-final String toolStableDartSdkVersion = '3.8.0';
-final String toolStableFlutterSdkVersion = '3.32.0';
+final String toolStableDartSdkVersion = '3.8.1';
+final String toolStableFlutterSdkVersion = '3.32.1';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);